server/requirements.txt
blinker==1.7.0
paddlenlp==2.8.1
albumentations==1.3.1
# paddlefsl==1.1.0
pandas==2.1.4
protobuf==3.20.3
PyYAML==6.0.1
urllib3==2.1.0
nvidia-pytriton==0.4.2
triton==2.1.0
tritonclient==2.40.0
# numpy==1.23.1
kafka-python==2.0.2
python-dotenv